requirements.txt
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-experimental
faiss-cpu
streamlit
pandas
pillow